Update README.md
Weather Or Not

An app using a weather api allowing users to check on the weather whilst simultaneously helping the user decide what may be appropriate to wear and style hair according to the their location and the daily weather expectations.

AAU should be able to get information on location weather for day. AAU should be able to see images that suggest what to wear or not to wear.

APP Screenshot

Future Capabilities Would like to connect to Amazon api to market products that user may be interested in once given suggested looks

Would like to implement a weekly live database update for new "looks" that stay current with daily trends.